+++ tests/testthat/test-algo-exp_smoothing-ets.R
@@ -1,4 +1,4 @@
-# ---- STANDARD ARIMA ----
+# ---- ETS, CROSTON ----
 context("TEST exp_smoothing: ets")
 
 
@@ -15,7 +15,7 @@ model_spec <- exp_smoothing() %>%
     set_engine("ets")
 
 
-# PARSNIP ----
+# ETS PARSNIP ----
 
 # * NO XREGS ----
 
@@ -71,12 +71,15 @@ test_that("exp_smoothing: ets, Test Predictions", {
 
 
 
-# ---- WORKFLOWS ----
+# ---- ETS WORKFLOWS ----
 
 # Model Spec
 model_spec <- exp_smoothing(
     seasonal_period = 12,
-    error = "multiplicative", trend = "additive", season = "multiplicative") %>%
+    error = "multiplicative", trend = "additive", season = "multiplicative"
+    ,
+    smooth_level = 0.2, smooth_trend = 0.1, smooth_seasonal = 0.1
+    ) %>%
     set_engine("ets")
 
 # Recipe spec
@@ -141,3 +144,71 @@ test_that("exp_smoothing: ets (workflow), Test Predictions", {
 
 })
 
+# ---- CROSTON WORKFLOWS ----
+
+# Model Spec
+model_spec <- exp_smoothing(
+    smooth_level = 0.2
+) %>%
+    set_engine("croston")
+
+# Recipe spec
+recipe_spec <- recipe(value ~ date, data = training(splits)) %>%
+    step_log(value, skip = FALSE)
+
+
+# Workflow
+wflw <- workflow() %>%
+    add_recipe(recipe_spec) %>%
+    add_model(model_spec)
+
+wflw_fit <- wflw %>%
+    fit(training(splits))
+
+# Forecast
+predictions_tbl <- wflw_fit %>%
+    modeltime_calibrate(testing(splits)) %>%
+    modeltime_forecast(new_data = testing(splits),
+                       actual_data = training(splits)) %>%
+    mutate_at(vars(.value), exp)
+
+# TESTS
+test_that("exp_smoothing: Arima (workflow), Test Model Fit Object", {
+
+    testthat::expect_s3_class(wflw_fit$fit$fit$fit, "croston_fit_impl")
+
+    # $fit
+
+    testthat::expect_s3_class(wflw_fit$fit$fit$fit$models$model_1, "forecast")
+
+    testthat::expect_s3_class(wflw_fit$fit$fit$fit$data, "tbl_df")
+
+    testthat::expect_equal(names(wflw_fit$fit$fit$fit$data)[1], "date")
+
+    testthat::expect_true(is.null(wflw_fit$fit$fit$fit$extras$xreg_recipe))
+
+    # $preproc
+    mld <- wflw_fit %>% workflows::pull_workflow_mold()
+    testthat::expect_equal(names(mld$outcomes), "value")
+
+})
+
+test_that("exp_smoothing: ets (workflow), Test Predictions", {
+
+    full_data <- bind_rows(training(splits), testing(splits))
+
+    # Structure
+    testthat::expect_identical(nrow(full_data), nrow(predictions_tbl))
+    testthat::expect_identical(full_data$date, predictions_tbl$.index)
+
+    # Out-of-Sample Accuracy Tests
+    predictions_tbl <- predictions_tbl %>% filter(.key == "prediction")
+    resid <- testing(splits)$value - predictions_tbl$.value
+
+    # - Max Error less than 1500
+    testthat::expect_lte(max(abs(resid)), 1500)
+
+    # - MAE less than 700
+    testthat::expect_lte(mean(abs(resid)), 1000)
+
+})